refactor: replace int sentinels with std::optional<size_t> throughout canvas
Eliminates all static_cast<int>/static_cast<size_t> round-trips and -1/0
sentinel comparisons. Node indices are now typed as std::optional<size_t>:
nullopt = no selection / root parent; value = valid index. Applies to
CanvasNode::parent_idx_, all five selection/drag/resize state members in
MindMapCanvasView, command classes, and branch drawing helpers.

+++ src/ui/branch/DrawBranchesBezier.cpp
@@ -15,7 +15,7 @@ constexpr ImU32 kColorEdge = IM_COL32(140, 170, 210, 255);  // NOLINT(hicpp-sign
 }  // namespace
 
 void DrawMindMapBranchBezier(
-    const BranchRenderContext& ctx, const int child_index,
+    const BranchRenderContext& ctx, const size_t child_index,
     const std::vector<mind_map::ui::CanvasNode>& nodes) {
   assert(ctx.draw_list_ != nullptr);
   BranchEdgeData g{};
